+++ test/DurableTask.SqlServer.Tests/Integration/DatabaseManagement.cs
@@ -1,5 +1,5 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
 
 namespace DurableTask.SqlServer.Tests.Integration
 {
@@ -41,7 +41,7 @@ namespace DurableTask.SqlServer.Tests.Integration
             var expectedScriptFiles = new HashSet<string>
             {
                 "drop-schema.sql",
-                "schema-0.2.0.sql",
+                "schema-1.0.0.sql",
                 "logic.sql",
                 "permissions.sql",
             };
@@ -92,7 +92,7 @@ namespace DurableTask.SqlServer.Tests.Integration
                 .Expect(
                     LogAssert.AcquiredAppLock(),
                     LogAssert.ExecutedSqlScript("drop-schema.sql"),
-                    LogAssert.ExecutedSqlScript("schema-0.2.0.sql"),
+                    LogAssert.ExecutedSqlScript("schema-1.0.0.sql"),
                     LogAssert.ExecutedSqlScript("logic.sql"),
                     LogAssert.ExecutedSqlScript("permissions.sql"),
                     LogAssert.SprocCompleted("dt._UpdateVersion"))
@@ -155,7 +155,7 @@ namespace DurableTask.SqlServer.Tests.Integration
                 .Expect(
                     LogAssert.AcquiredAppLock(),
                     LogAssert.SprocCompleted("dt._GetVersions"),
-                    LogAssert.ExecutedSqlScript("schema-0.2.0.sql"),
+                    LogAssert.ExecutedSqlScript("schema-1.0.0.sql"),
                     LogAssert.ExecutedSqlScript("logic.sql"),
                     LogAssert.ExecutedSqlScript("permissions.sql"),
                     LogAssert.SprocCompleted("dt._UpdateVersion"))
@@ -211,7 +211,7 @@ namespace DurableTask.SqlServer.Tests.Integration
                     // 1st
                     LogAssert.AcquiredAppLock(statusCode: 0),
                     LogAssert.SprocCompleted("dt._GetVersions"),
-                    LogAssert.ExecutedSqlScript("schema-0.2.0.sql"),
+                    LogAssert.ExecutedSqlScript("schema-1.0.0.sql"),
                     LogAssert.ExecutedSqlScript("logic.sql"),
                     LogAssert.ExecutedSqlScript("permissions.sql"),
                     LogAssert.SprocCompleted("dt._UpdateVersion"),
@@ -287,6 +287,8 @@ namespace DurableTask.SqlServer.Tests.Integration
                 "dt._RenewOrchestrationLocks",
                 "dt._RenewTaskLocks",
                 "dt._UpdateVersion",
+                "dt._RewindInstance",
+                "dt._RewindInstanceRecursive",
             };
 
             var expectedViewNames = new HashSet<string>(StringComparer.Ordinal)