+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## v1.0.0
+
+### New
+
+* Support for orchestration rewind ([#96](https://github.com/microsoft/durabletask-mssql/pull/96)) - contributed by [@Greybird](https://github.com/Greybird)
+
+### Updates
+
+* Updated package version to v1.0.0
+* Renamed schema-0.2.0.sql to schema-1.0.0.sql
+* Updated [Microsoft.Azure.WebJobs.Extensions.DurableTask](https://www.nuget.org/packages/Microsoft.Azure.WebJobs.Extensions.DurableTask) dependency to v2.7.*
+* Fixed data leak in ContinueAsNew path ([#102](https://github.com/microsoft/durabletask-mssql/pull/102))
+* Fixed inaccurate license headers
+
+### Breaking changes
+
+None
+
 ## v1.0.0-rc2
 
 ### New